# feat: optimize Spanner changestream metadata table