Adds indexes to the SpannerIO change streams metadata table. This serves to optimize two specific queries:

1. `getUnfinishedMinWatermark`
2. `getAllPartitionsCreatedAfter`

(micro-)Benchmarking results with a metadata table containing 11 million rows:

| Method | No Indexes (current) | Index (this PR) |
| --------- | -------- | ------- |
| getUnfinishedMinWatermark | 3.190s (+/- 0.91s) | 0.020s (+/- 0.002s) |
| getAllPartitionsCreatedAfter | 1.250s (+/- 0.230s) | 0.012s (+/- 0.001s) |


Customers can add these index manually if they want to by running the following DDL statements:

```sql
-- For GoogleSQL dialect
CREATE INDEX WatermarkIndex ON <Metadata Table Name> (Watermark) STORING (State);
CREATE INDEX CreatedAtStartTimestampIndex ON <Metadata Table Name> (CreatedAt, StartTimestamp);

-- For PostgreSQL dialect
CREATE INDEX "WatermarkIndex" ON <Metadata Table Name> ("Watermark") INCLUDE ("State");
CREATE INDEX "CreatedAtStartTimestampIndex" ON <Metadata Table Name> ("CreatedAt", "StartTimestamp");
```