@velamkao 2024-08-29
Hi @thiagotnunes would you know if this change to add the index is going to be part of Beam 2.59.0 release?