@velamkao 2024-08-29
> feat: index creation toga4/spream#106

Never mind, I can see its there https://github.com/apache/beam/blob/v2.59.0-RC1/sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/spanner/changestreams/dao/PartitionMetadataAdminDao.java